Histogram/requirements.txt
pip==20.2.3
setuptools==49.2.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
Pillow==8.4.0
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.4
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
